feat: extract asset-card out of search-item and used it in resources-options-pane
